Finished AStarSolver. Still need ReverseAStarSolver and both combined with AdvancedSolver.

+++ SnakeAI/HamiltonianBFSSolver.cpp
@@ -9,8 +9,8 @@ HamiltonianBFSSolver::HamiltonianBFSSolver(const SnakeState& gameState) :
 bool HamiltonianBFSSolver::uniaryOp(const SnakeState& gameState)
 {
 	// Calculate the number of empty spaces
-	int numCells = gameState.getNCells();
-	int nEmpty = numCells - gameState.getSnake().size();
+	int numCells = static_cast<int>(gameState.getNCells());
+	int nEmpty = numCells - static_cast<int>(gameState.getSnake().size());
 	
 	return (static_cast<float>(nEmpty) / numCells) < 0.5;
 }
